requirements.txt
Flask
pytesseract
Pillow
requests